third_party/WebKit/LayoutTests/images/element-gcd-while-generating-alt-content-expected.txt

CONSOLE ERROR: line 135: Uncaught HierarchyRequestError: Failed to execute 'insertNode' on 'Range': The node to be inserted contains the insertion point; it may not be inserted into itself.
